Fixed welcome.cr
I fixed it by using a channel_id instead of a guild_id at creating the welcome message. I also added the ability to set the channel where the welcome message should get sended to with !setwelcome (channel).

I know that this bot is still improvable (for example that he doesn't takes everything as an argument at !setwelcome) but I want it to be simple and not too complex.